fix last chunking errors with calendar
we imported some files that called assertMainOrNode() in the common date
chunk. size- and color related code moved to CalendarGuiUtils
