+++ src/calendar/view/CalendarEventBubble.ts
@@ -1,9 +1,9 @@
 import m, { Child, Children, Component, Vnode } from "mithril"
-import { colorForBg } from "../date/CalendarUtils"
 import { px, size } from "../../gui/size"
 import { Icon } from "../../gui/base/Icon"
 import { Icons } from "../../gui/base/icons/Icons"
 import type { clickHandler } from "../../gui/base/GuiUtils"
+import { colorForBg } from "../gui/CalendarGuiUtils.js"
 
 export type CalendarEventBubbleAttrs = {
 	text: string
