+++ src/calendar/view/CalendarView.ts
@@ -11,7 +11,7 @@ import type { CalendarEvent, GroupSettings, UserSettingsGroupRoot } from "../../
 import { createGroupSettings } from "../../api/entities/tutanota/TypeRefs.js"
 import { defaultCalendarColor, GroupType, Keys, reverse, ShareCapability, TimeFormat, WeekStart } from "../../api/common/TutanotaConstants"
 import { locator } from "../../api/main/MainLocator"
-import { getGroupColors, getStartOfTheWeekOffset, getStartOfTheWeekOffsetForUser, getTimeZone } from "../date/CalendarUtils"
+import { getStartOfTheWeekOffset, getStartOfTheWeekOffsetForUser, getTimeZone } from "../date/CalendarUtils"
 import { ButtonColor, ButtonType } from "../../gui/base/Button.js"
 import { CalendarMonthView } from "./CalendarMonthView"
 import { DateTime } from "luxon"
@@ -34,7 +34,7 @@ import { GroupInvitationFolderRow } from "../../sharing/view/GroupInvitationFold
 import { SidebarSection } from "../../gui/SidebarSection"
 import type { HtmlSanitizer } from "../../misc/HtmlSanitizer"
 import { ProgrammingError } from "../../api/common/error/ProgrammingError"
-import { calendarNavConfiguration, CalendarViewType, shouldDefaultToAmPmTimeFormat } from "../gui/CalendarGuiUtils.js"
+import { calendarNavConfiguration, CalendarViewType, getGroupColors, shouldDefaultToAmPmTimeFormat } from "../gui/CalendarGuiUtils.js"
 import { CalendarViewModel, MouseOrPointerEvent } from "./CalendarViewModel"
 import { showNewCalendarEventEditDialog } from "../gui/eventeditor-view/CalendarEventEditDialog.js"
 import { CalendarEventPopup } from "../gui/eventpopup/CalendarEventPopup.js"
